split and moved files around
